Converted everything to es6
