+++ .bowerrc
@@ -1,5 +1,5 @@
 {
-  "directory": "./app/client/statics/",
+  "directory": "src/app/client/statics/bower_components",
   "analytics": false,
   "interactive": false
 }
