+++ .eslintrc
@@ -1,6 +1,7 @@
 {
   "env": {
     "browser": true,
+    "es6": true,
     "jasmine": true,
     "node": true
   },
@@ -9,6 +10,21 @@
     "array-bracket-spacing": [
       2,
       "never"
+    ],
+    "arrow-body-style": [
+      2,
+      "always"
+    ],
+    "arrow-parens": [
+      2,
+      "always"
+    ],
+    "arrow-spacing": [
+      2,
+      {
+        "before": true,
+        "after": true
+      }
     ],
     "block-spacing": [
       2,
@@ -77,6 +93,8 @@
     ],
     "newline-after-var": 2,
     "no-array-constructor": 2,
+    "no-arrow-condition": 2,
+    "no-const-assign": 2,
     "no-eval": 2,
     "no-extend-native": 2,
     "no-invalid-this": 2,
@@ -96,8 +114,13 @@
       2,
       "nofunc"
     ],
+    "no-var": 2,
     "no-with": 2,
     "object-curly-spacing": [
+      2,
+      "never"
+    ],
+    "object-shorthand": [
       2,
       "never"
     ],
@@ -114,6 +137,8 @@
       2,
       "never"
     ],
+    "prefer-const": 2,
+    "prefer-template": 1,
     "quotes": [
       2,
       "single"
