--- app/server/controllers/heartbeat/index.js
@@ -1,9 +1,0 @@
-var controller = {
-  get: get
-};
-
-module.exports = controller;
-
-function get() {
-  return {message: 'OK'};
-}
