--- app/server/controllers/root.js
@@ -1,14 +1,0 @@
-var os = require('os');
-
-var controller = {
-  get: get
-};
-
-module.exports = controller;
-
-function get() {
-  return {
-    message: 'OK',
-    hostname: os.hostname()
-  };
-}
