--- gulp/tasks/nodemon.js
@@ -1,7 +1,0 @@
-module.exports = function(gulp, plugins, cfg) {
-  gulp.task('nodemon', nodemon);
-
-  function nodemon() {
-    plugins.nodemon(cfg.nodemon);
-  }
-};
