+++ gulpfile.js
@@ -1,4 +1,4 @@
 var gulp = require('gulp');
 var plugins = require('gulp-load-plugins')();
 
-require('./gulp')(gulp, plugins);
+require('./build/gulp')(gulp, plugins);
