+++ src/app/server/controllers/heartbeat/index.js
@@ -1,0 +1,9 @@
+const controller = {
+  get: get
+};
+
+module.exports = controller;
+
+function get() {
+  return {message: 'OK'};
+}
