+++ src/app/server/controllers/root.js
@@ -1,0 +1,14 @@
+const os = require('os');
+
+const controller = {
+  get: get
+};
+
+module.exports = controller;
+
+function get() {
+  return {
+    message: 'OK',
+    hostname: os.hostname()
+  };
+}
