+++ src/app/server/middleware/heartbeat.js
@@ -1,0 +1,13 @@
+const express = require('express');
+
+module.exports = function(app) {
+  const router = express.Router();
+
+  router
+    .get('/', function(req, res) { // eslint-disable-line
+      return res.json({message: 'OK'});
+    })
+    ;
+
+  app.use('/heartbeat', router);
+};
